requirements.txt
gradio
google-generativeai==0.8.4
python-dotenv
Flask==3.1.0
streamlit==1.42.0
streamlit-chat==0.1.1
